+++ src/main/scala/org/janjaali/sprayjwt/algorithms/HS512.scala
@@ -1,0 +1,10 @@
+package org.janjaali.sprayjwt.algorithms
+
+/**
+  * Represents HS512 hashing algorithm.
+  */
+case object HS512 extends HmacAlgorithm("HS512") {
+
+  override val cryptoAlgName = "HMACSHA512"
+
+}