+++ src/main/scala/org/janjaali/sprayjwt/algorithms/RS256.scala
@@ -1,0 +1,10 @@
+package org.janjaali.sprayjwt.algorithms
+
+/**
+  * Represents RS256 hashing algorithm.
+  */
+case object RS256 extends RsaAlgorithm("RS256") {
+
+  override protected val cryptoAlgName = "SHA256withRSA"
+
+}