+++ src/main/scala/org/janjaali/sprayjwt/algorithms/RS384.scala
@@ -1,0 +1,10 @@
+package org.janjaali.sprayjwt.algorithms
+
+/**
+  * Represents RS384 hashing algorithm.
+  */
+case object RS384 extends RsaAlgorithm("RS384") {
+
+  override protected val cryptoAlgName = "SHA384withRSA"
+
+}