+++ src/main/scala/org/janjaali/sprayjwt/encoder/ByteEncoder.scala
@@ -1,0 +1,20 @@
+package org.janjaali.sprayjwt.encoder
+
+/**
+  * ByteEncoder utility class.
+  */
+private[sprayjwt] object ByteEncoder {
+
+  private val encodingCharset = "UTF-8"
+
+  /**
+    * Encodes text into a byte array used UTF-8 charset.
+    *
+    * @param text the text to encode
+    * @return encoded byte array
+    */
+  def getBytes(text: String): Array[Byte] = {
+    text.getBytes(encodingCharset)
+  }
+
+}