+++ src/main/scala/org/janjaali/sprayjwt/exceptions/InvalidJwtHeaderException.scala
@@ -1,0 +1,8 @@
+package org.janjaali.sprayjwt.exceptions
+
+/**
+  * Exception thrown for invalid JWT header.
+  *
+  * @param message detailed message
+  */
+class InvalidJwtHeaderException(message: String) extends Exception(message)