+++ src/main/scala/org/janjaali/sprayjwt/exceptions/InvalidSignatureException.scala
@@ -1,0 +1,6 @@
+package org.janjaali.sprayjwt.exceptions
+
+/**
+  * Exception thrown for invalid signed JWT's.
+  */
+class InvalidSignatureException extends Exception("Invalid signature")