+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
-## Version 1.0.0
+## Version 1.0.0 (2017-11-12)
 
-Initial Release
+### New features
+- Encode and decode JWT token
+- suppot hashing algorithms: 
+- - HS256
+- - HS384
+- - HS512
+- - RS256
+- - RS384
+- - RS512
+- - ES256
+- - ES384
+- - ES512